make render manifest extensible
move JS and WASM specific template stuff into plugins
