+++ lib/ChunkTemplate.js
@@ -4,8 +4,6 @@
 */
 "use strict";
 
-const ConcatSource = require("webpack-sources").ConcatSource;
-const Template = require("./Template");
 const Tapable = require("tapable").Tapable;
 const SyncWaterfallHook = require("tapable").SyncWaterfallHook;
 const SyncHook = require("tapable").SyncHook;
@@ -15,6 +13,7 @@ module.exports = class ChunkTemplate extends Tapable {
 		super();
 		this.outputOptions = outputOptions || {};
 		this.hooks = {
+			renderManifest: new SyncWaterfallHook(["result", "options"]),
 			modules: new SyncWaterfallHook(["source", "chunk", "moduleTemplate", "dependencyTemplates"]),
 			render: new SyncWaterfallHook(["source", "chunk", "moduleTemplate", "dependencyTemplates"]),
 			renderWithEntry: new SyncWaterfallHook(["source", "chunk"]),
@@ -24,55 +23,11 @@ module.exports = class ChunkTemplate extends Tapable {
 	}
 
 	getRenderManifest(options) {
-		const chunk = options.chunk;
-		const outputOptions = options.outputOptions;
-		const moduleTemplates = options.moduleTemplates;
-		const dependencyTemplates = options.dependencyTemplates;
-
 		const result = [];
 
-		const filenameTemplate = outputOptions.chunkFilename;
-
-		result.push({
-			render: () => this.renderJavascript(chunk, moduleTemplates.javascript, dependencyTemplates),
-			filenameTemplate,
-			pathOptions: {
-				chunk
-			},
-			identifier: `chunk${chunk.id}`,
-			hash: chunk.hash
-		});
-
-		for(const module of chunk.getModules().filter(m => m.type && m.type.startsWith("webassembly"))) {
-			const filenameTemplate = outputOptions.webassemblyModuleFilename;
-
-			result.push({
-				render: () => this.renderWebAssembly(module, moduleTemplates.webassembly, dependencyTemplates),
-				filenameTemplate,
-				pathOptions: {
-					module
-				},
-				identifier: `webassemblyModule${module.id}`,
-				hash: module.hash
-			});
-		}
+		this.hooks.renderManifest.call(result, options);
 
 		return result;
-	}
-
-	renderJavascript(chunk, moduleTemplate, dependencyTemplates) {
-		const moduleSources = Template.renderChunkModules(chunk, m => true, moduleTemplate, dependencyTemplates);
-		const core = this.hooks.modules.call(moduleSources, chunk, moduleTemplate, dependencyTemplates);
-		let source = this.hooks.render.call(core, chunk, moduleTemplate, dependencyTemplates);
-		if(chunk.hasEntryModule()) {
-			source = this.hooks.renderWithEntry.call(source, chunk);
-		}
-		chunk.rendered = true;
-		return new ConcatSource(source, ";");
-	}
-
-	renderWebAssembly(module, moduleTemplate, dependencyTemplates) {
-		return moduleTemplate.render(module, dependencyTemplates, {});
 	}
 
 	updateHash(hash) {
