+++ lib/MainTemplate.js
@@ -7,6 +7,7 @@
 const ConcatSource = require("webpack-sources").ConcatSource;
 const OriginalSource = require("webpack-sources").OriginalSource;
 const PrefixSource = require("webpack-sources").PrefixSource;
+const RawSource = require("webpack-sources").RawSource;
 const Template = require("./Template");
 const Tapable = require("tapable").Tapable;
 const SyncWaterfallHook = require("tapable").SyncWaterfallHook;
@@ -34,6 +35,7 @@ module.exports = class MainTemplate extends Tapable {
 		super();
 		this.outputOptions = outputOptions || {};
 		this.hooks = {
+			renderManifest: new SyncWaterfallHook(["result", "options"]),
 			modules: new SyncWaterfallHook(["modules", "chunk", "hash", "moduleTemplate", "dependencyTemplates"]),
 			moduleObj: new SyncWaterfallHook(["source", "chunk", "hash", "moduleIdExpression"]),
 			requireEnsure: new SyncWaterfallHook(["source", "chunk", "hash", "chunkIdExpression"]),
@@ -73,8 +75,7 @@ module.exports = class MainTemplate extends Tapable {
 			source.add("/******/ })\n");
 			source.add("/************************************************************************/\n");
 			source.add("/******/ (");
-			const modules = Template.renderChunkModules(chunk, () => true, moduleTemplate, dependencyTemplates, "/******/ ");
-			source.add(this.hooks.modules.call(modules, chunk, hash, moduleTemplate, dependencyTemplates));
+			source.add(this.hooks.modules.call(new RawSource(""), chunk, hash, moduleTemplate, dependencyTemplates));
 			source.add(")");
 			return source;
 		});
@@ -207,33 +208,9 @@ module.exports = class MainTemplate extends Tapable {
 	}
 
 	getRenderManifest(options) {
-		const chunk = options.chunk;
-		const hash = options.hash;
-		const fullHash = options.fullHash;
-		const outputOptions = options.outputOptions;
-		const moduleTemplates = options.moduleTemplates;
-		const dependencyTemplates = options.dependencyTemplates;
-
 		const result = [];
 
-		let filenameTemplate;
-		if(chunk.filenameTemplate)
-			filenameTemplate = chunk.filenameTemplate;
-		else
-			filenameTemplate = outputOptions.filename;
-
-		const useChunkHash = this.useChunkHash(chunk);
-
-		result.push({
-			render: () => this.render(hash, chunk, moduleTemplates.javascript, dependencyTemplates),
-			filenameTemplate,
-			pathOptions: {
-				noChunkHash: !useChunkHash,
-				chunk
-			},
-			identifier: `chunk${chunk.id}`,
-			hash: useChunkHash ? chunk.hash : fullHash
-		});
+		this.hooks.renderManifest.call(result, options);
 
 		return result;
 	}
