+++ lib/WebAssemblyModulesPlugin.js
@@ -17,7 +17,36 @@ class WebAssemblyModulesPlugin {
 			normalModuleFactory.hooks.createParser.for("webassembly/experimental").tap("WebAssemblyModulesPlugin", () => {
 				return new WebAssemblyParser();
 			});
+
+			compilation.chunkTemplate.hooks.renderManifest.tap("WebAssemblyModulesPlugin", (result, options) => {
+				const chunk = options.chunk;
+				const outputOptions = options.outputOptions;
+				const moduleTemplates = options.moduleTemplates;
+				const dependencyTemplates = options.dependencyTemplates;
+
+				for(const module of chunk.modulesIterable) {
+					if(module.type && module.type.startsWith("webassembly")) {
+						const filenameTemplate = outputOptions.webassemblyModuleFilename;
+
+						result.push({
+							render: () => this.renderWebAssembly(module, moduleTemplates.webassembly, dependencyTemplates),
+							filenameTemplate,
+							pathOptions: {
+								module
+							},
+							identifier: `webassemblyModule${module.id}`,
+							hash: module.hash
+						});
+					}
+				}
+
+				return result;
+			});
 		});
+	}
+
+	renderWebAssembly(module, moduleTemplate, dependencyTemplates) {
+		return moduleTemplate.render(module, dependencyTemplates, {});
 	}
 }
